// src/app/models/material.model.ts
export interface Material {
    id: number;
    material: string;
    id_material_type: number;
    is_default: boolean;
    coverage_sq: number;
    converage_lf: number;
    id_unit_buy: number;
    id_trademark: number;
    id_supplier_default: number;
    is_active: number;
    created_by: string;
}